split packet_stream module into multiple files
