--- src/packet_stream.rs
@@ -1,163 +1,0 @@
-use std::pin::Pin;
-
-use futures::{
-    future::Future,
-    io::{self, AsyncRead, AsyncReadExt, AsyncWrite, AsyncWriteExt},
-};
-use tokio_util::compat::{Compat, TokioAsyncWriteCompatExt};
-use tun::{DeviceReader, DeviceWriter};
-
-use crate::common::{AsyncReadFixed, AsyncWriteFixed};
-
-pub trait PacketReceiver: Send {
-    fn receive(&mut self) -> impl Future<Output = io::Result<Box<[u8]>>> + Send;
-}
-
-// unused now, but may be useful later
-#[allow(dead_code)]
-pub trait DynPacketReceiver: Send {
-    fn receive_dyn(&mut self) -> Pin<Box<dyn Future<Output = io::Result<Box<[u8]>>> + Send + '_>>;
-}
-
-impl<R: PacketReceiver> DynPacketReceiver for R {
-    fn receive_dyn(&mut self) -> Pin<Box<dyn Future<Output = io::Result<Box<[u8]>>> + Send + '_>> {
-        Box::pin(self.receive())
-    }
-}
-
-pub struct TaggedPacketReceiver<IO: Send> {
-    stream: IO,
-}
-
-impl<IO: AsyncRead + Unpin + Send> TaggedPacketReceiver<IO> {
-    pub fn new(stream: IO) -> Self {
-        Self { stream }
-    }
-}
-
-impl<IO: AsyncRead + Unpin + Send> PacketReceiver for TaggedPacketReceiver<IO> {
-    async fn receive(&mut self) -> io::Result<Box<[u8]>> {
-        let packet_size = self.stream.read_u16().await? as usize;
-        let mut packet = vec![0u8; packet_size].into_boxed_slice();
-
-        let mut offset = 0;
-        while offset < packet_size {
-            let received = self.stream.read(&mut packet[offset..]).await?;
-            if received == 0 {
-                return Err(io::ErrorKind::UnexpectedEof.into());
-            }
-            offset += received;
-        }
-
-        Ok(packet)
-    }
-}
-
-pub struct TunReceiver {
-    reader: DeviceReader,
-    buffer: Vec<u8>,
-}
-
-impl TunReceiver {
-    pub fn new(reader: DeviceReader, mtu: usize) -> Self {
-        Self {
-            reader,
-            buffer: vec![0; mtu],
-        }
-    }
-}
-
-impl PacketReceiver for TunReceiver {
-    async fn receive(&mut self) -> io::Result<Box<[u8]>> {
-        // this is not cancel-safe, but we do not particularly care
-        let cnt_read =
-            <DeviceReader as tokio::io::AsyncReadExt>::read(&mut self.reader, &mut self.buffer)
-                .await?;
-        Ok(self.buffer[..cnt_read].into())
-    }
-}
-
-pub trait PacketSender: Send {
-    fn send(&mut self, packet: &[u8]) -> impl Future<Output = io::Result<()>> + Send;
-
-    fn close(&mut self) -> impl Future<Output = io::Result<()>> + Send;
-}
-
-pub trait DynPacketSender: Send {
-    fn send_dyn<'a>(
-        &'a mut self,
-        packet: &'a [u8],
-    ) -> Pin<Box<dyn Future<Output = io::Result<()>> + Send + 'a>>;
-
-    fn close_dyn(&mut self) -> Pin<Box<dyn Future<Output = io::Result<()>> + Send + '_>>;
-}
-
-impl<S: PacketSender> DynPacketSender for S {
-    fn send_dyn<'a>(
-        &'a mut self,
-        packet: &'a [u8],
-    ) -> Pin<Box<dyn Future<Output = io::Result<()>> + Send + 'a>> {
-        Box::pin(self.send(packet))
-    }
-
-    fn close_dyn(&mut self) -> Pin<Box<dyn Future<Output = io::Result<()>> + Send + '_>> {
-        Box::pin(self.close())
-    }
-}
-
-pub struct TaggedPacketSender<IO> {
-    stream: IO,
-}
-
-impl<IO: AsyncWrite + Unpin> TaggedPacketSender<IO> {
-    pub fn new(stream: IO) -> Self {
-        Self { stream }
-    }
-}
-
-impl<IO: AsyncWrite + Unpin + Send> PacketSender for TaggedPacketSender<IO> {
-    async fn send(&mut self, packet: &[u8]) -> io::Result<()> {
-        let packet_size = match u16::try_from(packet.len()) {
-            Ok(s) => s,
-            Err(_) => return Err(io::ErrorKind::FileTooLarge.into()),
-        };
-        self.stream.write_u16(packet_size).await?;
-
-        let mut offset = 0;
-        while offset < packet.len() {
-            let written = self.stream.write(&packet[offset..]).await?;
-            if written == 0 {
-                return Err(io::ErrorKind::WriteZero.into());
-            }
-            offset += written;
-        }
-        self.stream.flush().await
-    }
-
-    async fn close(&mut self) -> io::Result<()> {
-        self.stream.close().await
-    }
-}
-
-pub struct TunSender {
-    wrapped: Compat<DeviceWriter>,
-}
-
-impl From<DeviceWriter> for TunSender {
-    fn from(value: DeviceWriter) -> Self {
-        Self {
-            wrapped: value.compat_write(),
-        }
-    }
-}
-
-impl PacketSender for TunSender {
-    async fn send(&mut self, packet: &[u8]) -> io::Result<()> {
-        self.wrapped.write_all(packet).await?;
-        self.wrapped.flush().await
-    }
-
-    async fn close(&mut self) -> io::Result<()> {
-        self.wrapped.close().await
-    }
-}
